Separate files and improve readme
