+++ src/challenges/js/30DaysJs-30/playground.json
@@ -3,5 +3,13 @@
   "showFileExplorer": false,
   "showPreview": true,
   "template": "vanilla-js",
-  "runner": "sandpack"
+  "runner": "sandpack",
+  "entryFile": "exercise.js",
+  "excludeFiles": [],
+  "hideFiles": [
+    ".*.test.js$",
+    "styles.css$",
+    "index.js$",
+    "index.html$"
+  ]
 }
