+++ src/challenges/js/30DaysJs-30/src/User.js
@@ -1,0 +1,17 @@
+export class User {
+  constructor(name) {
+    // Tu código aquí 👈
+  }
+
+  login() {
+    // Tu código aquí 👈
+  }
+
+  logout() {
+    // Tu código aquí 👈
+  }
+
+  isLoggedIn() {
+    // Tu código aquí 👈
+  }
+}
